Refactor Program Manager to remove resolver and split record resolution to a separate object
