+++ rust/src/program/mod.rs
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with the Aleo library. If not, see <https://www.gnu.org/licenses/>.
 
-use crate::api::{config::NetworkConfig, AleoAPIClient};
+use crate::AleoAPIClient;
 use snarkvm_console::{
     account::PrivateKey,
     program::{Ciphertext, Network, ProgramID},
@@ -23,11 +23,7 @@ use snarkvm_synthesizer::Program;
 
 use anyhow::{anyhow, bail, Result};
 use indexmap::IndexMap;
-use snarkvm_console::program::Address;
 use std::path::PathBuf;
-
-pub mod config;
-pub use config::*;
 
 pub mod helpers;
 pub use helpers::*;
@@ -35,10 +31,11 @@ pub use helpers::*;
 pub mod network;
 pub use network::*;
 
-pub mod resolvers;
-pub use resolvers::*;
+pub mod resolver;
+pub use resolver::*;
 
 pub mod transfer;
+use crate::Encryptor;
 pub use transfer::*;
 
 /// Program management object for loading programs for building, execution, and deployment
@@ -47,80 +44,83 @@ pub use transfer::*;
 /// CLI tools, IDE plugins, Server-side stack components and other software that needs to
 /// interact with the Aleo network.
 #[derive(Clone)]
-pub struct ProgramManager<N: Network, R: Resolver<N>> {
+pub struct ProgramManager<N: Network> {
     pub(crate) programs: IndexMap<ProgramID<N>, Program<N>>,
     pub(crate) private_key: Option<PrivateKey<N>>,
     pub(crate) private_key_ciphertext: Option<Ciphertext<N>>,
+    pub(crate) local_program_directory: Option<PathBuf>,
     pub(crate) api_client: Option<AleoAPIClient<N>>,
-    pub(crate) resolver: R,
 }
 
-impl<N: Network, R: Resolver<N>> ProgramManager<N, R> {
+impl<N: Network> ProgramManager<N> {
     /// Create a new program manager by specifying custom options for the private key (or private
     /// key ciphertext) and resolver. Use this method if you want to create a custom resolver
     /// (i.e. one that searches a local or remote database) for program and record resolution.
     pub fn new(
         private_key: Option<PrivateKey<N>>,
         private_key_ciphertext: Option<Ciphertext<N>>,
-        network_config: Option<NetworkConfig>,
-        resolver: R,
+        api_client: Option<AleoAPIClient<N>>,
+        local_program_directory: Option<PathBuf>,
     ) -> Result<Self> {
         if private_key.is_some() && private_key_ciphertext.is_some() {
             bail!("Cannot have both private key and private key ciphertext");
         } else if private_key.is_none() && private_key_ciphertext.is_none() {
             bail!("Must have either private key or private key ciphertext");
         }
-        let api_client = network_config.map(|config| AleoAPIClient::<N>::from(&config));
         let programs = IndexMap::new();
-        Ok(Self { programs, private_key, private_key_ciphertext, api_client, resolver })
+        Ok(Self { programs, private_key, private_key_ciphertext, local_program_directory, api_client })
     }
 
-    /// Create a program manager with a file resolver that only uses the local file system to
-    /// resolve programs and records
-    pub fn program_manager_with_local_resource_resolution(
-        private_key: Option<PrivateKey<N>>,
-        private_key_ciphertext: Option<Ciphertext<N>>,
-        address: &Address<N>,
-        local_directory: impl TryInto<PathBuf>,
-        network_config: Option<NetworkConfig>,
-    ) -> Result<ProgramManager<N, FileSystemResolver<N>>> {
-        let local_directory = local_directory.try_into().map_err(|_| anyhow!("Path specified was not valid"))?;
-        let resolver = FileSystemResolver::new(&local_directory, Some(*address))?;
-        ProgramManager::<N, FileSystemResolver<N>>::new(private_key, private_key_ciphertext, network_config, resolver)
+    /// Manually add a program to the program manager from memory if it does not already exist
+    pub fn add_program(&mut self, program: &Program<N>) -> Result<()> {
+        if self.contains_program(program.id())? {
+            bail!("program already exists")
+        };
+        self.programs.entry(*program.id()).or_insert(program.clone());
+        Ok(())
     }
 
-    /// Create a program manager configured with a network resolver which only uses the Aleo network
-    /// to resolve programs and records
-    pub fn program_manager_with_network_resolution(
-        private_key: Option<PrivateKey<N>>,
-        private_key_ciphertext: Option<Ciphertext<N>>,
-        network_config: NetworkConfig,
-    ) -> Result<ProgramManager<N, AleoNetworkResolver<N>>> {
-        let resolver = AleoNetworkResolver::new(&network_config);
-        ProgramManager::<N, AleoNetworkResolver<N>>::new(
-            private_key,
-            private_key_ciphertext,
-            Some(network_config),
-            resolver,
-        )
+    /// Manually add a program to the program manager if it does not already exist or update
+    /// it if it does
+    pub fn update_program(&mut self, program: &Program<N>) -> Option<Program<N>> {
+        self.programs.insert(*program.id(), program.clone())
     }
 
-    /// Create a hybrid program manager which has the following behavior
-    ///
-    /// 1. Programs will first be searched for on the local file system, if a program or its imports
-    /// cannot be found within the specified directory, the resolver will attempt to resolve the
-    /// missing programs from the Aleo network
-    /// 2. The resolver will always attempt to resolve records on the Aleo network (and not within
-    /// the local file system)
-    pub fn program_manager_with_hybrid_resolution(
-        private_key: Option<PrivateKey<N>>,
-        private_key_ciphertext: Option<Ciphertext<N>>,
-        local_directory: impl TryInto<PathBuf>,
-        network_config: NetworkConfig,
-    ) -> Result<ProgramManager<N, HybridResolver<N>>> {
-        let local_directory = local_directory.try_into().map_err(|_| anyhow!("Path specified was not valid"))?;
-        let resolver = HybridResolver::new(&network_config, &local_directory)?;
-        ProgramManager::<N, HybridResolver<N>>::new(private_key, private_key_ciphertext, Some(network_config), resolver)
+    /// Retrieve a program from the program manager if it exists
+    pub fn get_program(&self, program_id: impl TryInto<ProgramID<N>>) -> Result<Program<N>> {
+        let program_id = program_id.try_into().map_err(|_| anyhow!("invalid program id"))?;
+        self.programs.get(&program_id).map_or(Err(anyhow!("program not found")), |program| Ok(program.clone()))
+    }
+
+    /// Determine if a program exists in the program manager
+    pub fn contains_program(&self, program_id: impl TryInto<ProgramID<N>>) -> Result<bool> {
+        let program_id = program_id.try_into().map_err(|_| anyhow!("invalid program id"))?;
+        Ok(self.programs.contains_key(&program_id))
+    }
+
+    /// Get the private key from the program manager. If the key is stored as ciphertext, a
+    /// password must be provided to decrypt it
+    pub(super) fn get_private_key(&self, password: Option<&str>) -> Result<PrivateKey<N>> {
+        if self.private_key.is_none() && self.private_key_ciphertext.is_none() {
+            bail!("Private key is not configured");
+        };
+        if let Some(private_key) = &self.private_key {
+            if self.private_key_ciphertext.is_some() {
+                bail!(
+                    "Private key ciphertext is also configured, cannot have both private key and private key ciphertext"
+                );
+            }
+            return Ok(*private_key);
+        };
+        if let Some(ciphertext) = &self.private_key_ciphertext {
+            if self.private_key.is_some() {
+                bail!("Private key is already configured, cannot have both private key and private key ciphertext");
+            }
+
+            let password = password.ok_or_else(|| anyhow!("Private key is encrypted, password is required"))?;
+            return Encryptor::<N>::decrypt_private_key_with_secret(ciphertext, password);
+        };
+        bail!("Private key configuration error")
     }
 }
 
@@ -128,79 +128,119 @@ impl<N: Network, R: Resolver<N>> ProgramManager<N, R> {
 mod tests {
 
     use super::*;
-    use crate::{Encryptor, RECIPIENT_PRIVATE_KEY};
+    use crate::{
+        test_utils::{HELLO_PROGRAM, HELLO_PROGRAM_2},
+        Encryptor,
+        RECIPIENT_PRIVATE_KEY,
+    };
     use snarkvm_console::network::Testnet3;
 
     use std::str::FromStr;
 
     #[test]
     fn test_constructors_fail_with_multiple_keys_or_no_keys() {
-        let network_config = NetworkConfig::testnet3();
+        let api_client = AleoAPIClient::<Testnet3>::testnet3();
         let private_key = PrivateKey::<Testnet3>::from_str(RECIPIENT_PRIVATE_KEY).unwrap();
-        let address = Address::<Testnet3>::try_from(&private_key).unwrap();
         let private_key_ciphertext =
             Encryptor::<Testnet3>::encrypt_private_key_with_secret(&private_key, "password").unwrap();
         // Create a temp dir without proper programs to test that the hybrid client works even if the local resource directory doesn't exist
         let temp_dir = std::env::temp_dir();
 
-        let file_program_manager =
-            ProgramManager::<Testnet3, FileSystemResolver<Testnet3>>::program_manager_with_local_resource_resolution(
-                None,
-                None,
-                &address,
-                temp_dir.clone(),
-                Some(network_config.clone()),
-            );
+        // Ensure that program manager creation fails if no key is provided
+        let program_manager =
+            ProgramManager::<Testnet3>::new(None, None, Some(api_client.clone()), Some(temp_dir.clone()));
 
-        assert!(file_program_manager.is_err());
+        assert!(program_manager.is_err());
 
-        let hybrid_program_manager =
-            ProgramManager::<Testnet3, HybridResolver<Testnet3>>::program_manager_with_hybrid_resolution(
-                None,
-                None,
-                temp_dir.clone(),
-                network_config.clone(),
-            );
+        // Ensure that program manager creation fails if both key and key ciphertext are provided
+        let program_manager = ProgramManager::<Testnet3>::new(
+            Some(private_key),
+            Some(private_key_ciphertext.clone()),
+            Some(api_client.clone()),
+            Some(temp_dir.clone()),
+        );
 
-        assert!(hybrid_program_manager.is_err());
+        assert!(program_manager.is_err());
 
-        let network_program_manager =
-            ProgramManager::<Testnet3, AleoNetworkResolver<Testnet3>>::program_manager_with_network_resolution(
-                None,
-                None,
-                network_config.clone(),
-            );
+        // Ensure program manager is created successfully if only a private key is provided
+        let program_manager =
+            ProgramManager::<Testnet3>::new(Some(private_key), None, Some(api_client.clone()), Some(temp_dir.clone()));
 
-        assert!(network_program_manager.is_err());
+        assert!(program_manager.is_ok());
 
-        let file_program_manager =
-            ProgramManager::<Testnet3, FileSystemResolver<Testnet3>>::program_manager_with_local_resource_resolution(
-                Some(private_key.clone()),
-                Some(private_key_ciphertext.clone()),
-                &address,
-                temp_dir.clone(),
-                Some(network_config.clone()),
-            );
+        // Ensure program manager is created successfully if only a private key ciphertext is provided
+        let program_manager =
+            ProgramManager::<Testnet3>::new(None, Some(private_key_ciphertext), Some(api_client), Some(temp_dir));
 
-        assert!(file_program_manager.is_err());
+        assert!(program_manager.is_ok());
+    }
 
-        let hybrid_program_manager =
-            ProgramManager::<Testnet3, HybridResolver<Testnet3>>::program_manager_with_hybrid_resolution(
-                Some(private_key.clone()),
-                Some(private_key_ciphertext.clone()),
-                temp_dir.clone(),
-                network_config.clone(),
-            );
+    #[test]
+    fn test_program_management_methods() {
+        let private_key = PrivateKey::<Testnet3>::from_str(RECIPIENT_PRIVATE_KEY).unwrap();
+        let mut program_manager = ProgramManager::<Testnet3>::new(Some(private_key), None, None, None).unwrap();
 
-        assert!(hybrid_program_manager.is_err());
+        // Test program addition
+        let program = Program::<Testnet3>::from_str(HELLO_PROGRAM).unwrap();
+        assert!(!program_manager.contains_program(program.id()).unwrap());
+        program_manager.add_program(&program).unwrap();
+        assert!(program_manager.contains_program(program.id()).unwrap());
+        assert_eq!(program_manager.get_program(program.id()).unwrap(), program);
+        assert_eq!(program_manager.get_program("hello.aleo").unwrap(), program);
+        assert!(program_manager.contains_program("hello.aleo").unwrap());
+        assert!(program_manager.add_program(&program).is_err());
 
-        let network_program_manager =
-            ProgramManager::<Testnet3, AleoNetworkResolver<Testnet3>>::program_manager_with_network_resolution(
-                Some(private_key.clone()),
-                Some(private_key_ciphertext.clone()),
-                network_config.clone(),
-            );
+        // Test program update methods
+        let program_2 = Program::<Testnet3>::from_str(HELLO_PROGRAM_2).unwrap();
+        let replaced_program = program_manager.update_program(&program_2).unwrap();
+        let retrieved_program = program_manager.get_program(program.id()).unwrap();
+        assert_eq!(replaced_program, program);
+        assert_eq!(retrieved_program, program_2);
+    }
 
-        assert!(network_program_manager.is_err());
+    #[test]
+    fn test_private_key_retrieval_from_ciphertext() {
+        let private_key = PrivateKey::from_str(RECIPIENT_PRIVATE_KEY).unwrap();
+        let private_key_ciphertext =
+            Encryptor::<Testnet3>::encrypt_private_key_with_secret(&private_key, "password").unwrap();
+        let temp_dir = std::env::temp_dir();
+        let api_client = AleoAPIClient::<Testnet3>::testnet3();
+
+        let program_manager = ProgramManager::<Testnet3>::new(
+            None,
+            Some(private_key_ciphertext),
+            Some(api_client),
+            Some(temp_dir.clone()),
+        )
+        .unwrap();
+
+        // Assert private key recovers correctly
+        let recovered_private_key = program_manager.get_private_key(Some("password")).unwrap();
+        assert_eq!(recovered_private_key, private_key);
+
+        // Assert error is thrown if password is incorrect
+        let recovered_private_key = program_manager.get_private_key(Some("wrong_password"));
+        assert!(recovered_private_key.is_err());
+
+        // Assert error is thrown if password is not provided
+        let recoverd_private_key = program_manager.get_private_key(None);
+        assert!(recoverd_private_key.is_err());
+    }
+
+    #[test]
+    fn test_private_key_retrieval_from_plaintext() {
+        let private_key = PrivateKey::<Testnet3>::from_str(RECIPIENT_PRIVATE_KEY).unwrap();
+        let temp_dir = std::env::temp_dir();
+        let api_client = AleoAPIClient::<Testnet3>::testnet3();
+
+        let program_manager =
+            ProgramManager::<Testnet3>::new(Some(private_key), None, Some(api_client), Some(temp_dir.clone())).unwrap();
+
+        // Assert private key recovers correctly regardless of password
+        let recovered_private_key = program_manager.get_private_key(None).unwrap();
+        assert_eq!(recovered_private_key, private_key);
+
+        let recovered_private_key = program_manager.get_private_key(Some("password")).unwrap();
+        assert_eq!(recovered_private_key, private_key);
     }
 }
